Better business/ui separation.
