+++ lib/src/user_exception.dart
@@ -1,5 +1,6 @@
 import 'dart:ui';
 
+import 'package:async_redux/src/store_provider.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
