+++ test/model_observer_test.dart
@@ -1,6 +1,5 @@
-import 'dart:async';
-
 import 'package:async_redux/async_redux.dart';
+import 'package:async_redux/src/store_provider.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
